+++ src/constraint_handler/PropagatorVariables.py
@@ -85,6 +85,16 @@ class VariableManager:
 
     def __init__(self):
         self.variables: dict[clingo.Symbol, dict[VariableTypeNames, VariableType]] = {}
+        self.user_variable_names: set[clingo.Symbol] = set()
+    
+    def add_user_variable_name(self, name: clingo.Symbol) -> None:
+        """
+        Add a variable name to the set of user variable names.
+
+        Args:
+            name: The variable name to add.
+        """
+        self.user_variable_names.add(name)
 
     def add_variable(self, name: clingo.Symbol, variable: VariableType) -> None:
         """
@@ -97,6 +107,17 @@ class VariableManager:
         if name not in self.variables:
             self.variables[name] = {}
         self.variables[name][variable.__class__.__name__] = variable
+
+    def is_user_variable(self, name: clingo.Symbol) -> bool:
+        """
+        Check if a variable name is a user variable.
+
+        Args:
+            name: The variable name to check.
+        Returns:
+            bool: True if the variable name is a user variable, False otherwise.
+        """
+        return name in self.user_variable_names
 
     def get_variables(self) -> Iterable[VariableType]:
         """