+++ src/constraint_handler/schemas/propagator_atom.py
@@ -1,5 +1,8 @@
 from __future__ import annotations
 
+from typing import NamedTuple
+
+import constraint_handler.schemas.expression as expression
 import constraint_handler.schemas.atom as atom
 import constraint_handler.schemas.domain as domain
 import constraint_handler.schemas.warning as warning
@@ -86,3 +89,8 @@ class Propagator_warning_forbid(warning.Warning_forbid):
 
 class Propagator_warning_ignore(warning.Warning_ignore):
     pass
+
+
+class Propagator_variable_interface(NamedTuple):
+    label: expression.constant
+    variable: expression.constant